feat: Adds add_day_time_cols method to SSWeather
add_day_time_cols is a method that creates the 'Day' and 'Time' columns ina weather file when the dataframe is indexed by datetime.
This is useful when we create new weather file from scratch. The dataframe is often index by DateTime and the 'Day' and 'Time' columns are added before writing to file.