feat: transferred solcast_historic
Added a script to create a weather file from SolCast historic data. Transferred from a script from v1.0.a and this is the first stab of breaking it down and refactoring it.

Future work:
To further break down the scripts into separate functions to help scalability and testing.
Add tests for this.